Improve cue sheet playlist playback.
No matter what playback position had been selected, the displayed
title was continually the title of the currently played track. It
did not change when selecting a playback position before or after
the current track.

Check the playback position. If inside the currently played track,
display the track title, If not, display the file's name.