Enable skins to leave current volume unchanged at GUI startup.
Starting with svn r37055, the meaning of zero as a potentiometer's
"default" parameter has changed.

Prior to this revision the volume set before the GUI's startup
remained unchanged with such default (due to a bug), after that
the volume was consequently set to zero.

In order to re-enable the old behavior, i.e. leaving the volume
unchanged at GUI startup, a plain hyphen-minus can be used as
"default" parameter now.

Additionally, update the documentation.