configure: Fix detection for inet_aton function
Unbreak the build with internal FFmpeg. The misdetection leads to a
build failure with modern FFmpeg because in that case FFmpeg defines
a static function inet_aton itself.

The program source to test for inet_aton runs into a compile error
because the missing declaration of inet_aton that is treated as
an error by at least GCC with the given CFLAGS. Fix that by
defining _BSD_SOURCE before inclusion of inet/arpa.h .

The headers here indicate that _BSD_SOURCE is deprecated. So fixing
this in configure and/or at least in the MPlayer source files,
that actually use inet_aton, would be a good thing. Those would be:

  stream/librtsp/rtsp_rtp.c
  stream/tcp.c
  stream/udp.c
  udp_sync.c

Maybe like it was done in FFmpeg commit af1818276ef271af98 .