demux_mng,vo_mng: Avoid inclusion of jpeg.h
If libmng is compiled with JNG support, it pulls in jpeg.h from
within its public headers. To my knowledge we do not use JNG
directly. This change does not disable the functionality to read
JNG files with demuxer mng.

Primary this helps compilation on current Debian. The header stdio.h
needs to be included before jpeg.h , but that #include is patched out
in Debian's libmng package.

Maybe we should revert this change after Debian (and derived)
packages are fixed.